refactor: 수익률 계산 로직 로또통계로 이동
