--- src/main/java/lotto/domain/ProfitCalculator.java
@@ -1,18 +1,0 @@
-package lotto.domain;
-
-import java.util.List;
-
-public class ProfitCalculator {
-    public static float calculateProfitRate(int totalPurchaseAmount, List<Integer> equalNumberCountList) {
-        int totalProfitAmount = calculateTotalProfit(equalNumberCountList);
-
-        float profitRate = (float) totalProfitAmount / totalPurchaseAmount;
-        return Math.round(profitRate * 100) / 100.0f;
-    }
-
-    public static int calculateTotalProfit(List<Integer> equalNumberCountList) {
-        return equalNumberCountList.stream()
-                .mapToInt(LottoRank::findPrizeMoneyFromEqualNumberCount)
-                .sum();
-    }
-}
